pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sanctum-ai"
version = "0.1.0"
description = "Python SDK for SanctumAI — secure credential management for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "SanctumAI", email = "hello@sanctumai.dev" }]
keywords = ["sanctum", "ai", "credentials", "secrets", "vault", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = ["pynacl>=1.5.0"]

[project.urls]
Homepage = "https://github.com/jwgale/sanctum"
Repository = "https://github.com/jwgale/sanctum-sdk-python"
Issues = "https://github.com/jwgale/sanctum-sdk-python/issues"
